rename: 비밀번호 재설정 파일명 변경(api와 동일하게)
