+++ src/components/auth/PasswordReset/PasswordResetForm.tsx
@@ -3,7 +3,7 @@
 import React from "react";
 import TextInput from "@/components/common/Input/TextInput";
 import BaseButton from "@/components/common/Button/BaseButton";
-import useResetPasswordForm from "./useResetPasswordForm";
+import useResetPasswordForm from "./usePasswordResetForm";
 import { EvenIcon } from "@/components/common/Icons";
 import Link from "next/link";
